Add conditional logic to migration

+++ db/migrate/20160725185946_remove_bank_account_id_from_expense_receipts.rb
@@ -1,5 +1,7 @@
 class RemoveBankAccountIdFromExpenseReceipts < ActiveRecord::Migration
   def change
-    remove_column :expense_receipts, :bank_account_id
+    if column_exists?(:expense_receipts, :bank_account_id)
+      remove_column :expense_receipts, :bank_account_id
+    end
   end
 end
